+++ packages/article-magazine-comment/__tests__/mocks.web.js
@@ -1,6 +1,5 @@
 // eslint-disable-next-line global-require
 jest.mock("@times-components/ad", () => require("./ad-mock"));
-jest.mock("@times-components/article", () => "Article");
 jest.mock("@times-components/article-byline", () => ({
   ArticleBylineWithLinks: "ArticleBylineWithLinks"
 }));
