+++ packages/article-magazine-comment/__tests__/shared-props.js
@@ -1,0 +1,19 @@
+import { adConfig } from "./ad-mock";
+
+const sharedProps = {
+  adConfig,
+  analyticsStream: () => {},
+  onAuthorPress: () => {},
+  onCommentGuidelinesPress: () => {},
+  onCommentsPress: () => {},
+  onLinkPress: () => {},
+  onRelatedArticlePress: () => {},
+  onTopicPress: () => {},
+  onTwitterLinkPress: () => {},
+  onVideoPress: () => {},
+  onViewed: () => {},
+  receiveChildList: () => {},
+  refetch: () => {}
+};
+
+export default sharedProps;
